当内容为空，且存在配置文件时，读取配置文件内容

+++ bin/born.py
@@ -14,8 +14,11 @@ def readFile(fname):
 
 def jsonObj(jsonFile):
 	jsonContent = readFile(jsonFile)
-	robj = json.loads(jsonContent,strict = False)
-	return robj 
+	try:
+		robj = json.loads(jsonContent,strict = False)
+		return robj
+	except:
+		return {} 
 
 def renderJson(tplContent,jsonObj):
 	return tplContent % jsonObj
@@ -24,14 +27,26 @@ def simpleReplace(content):
 	m = re.compile("\$([a-zA-Z]+)(\s?)")
 	return re.sub(m,"%(\\1)s\\2",content)
 
+def isFileEmpty(fileName):
+	f = open(fileName,"r")
+	c = f.read()
+	return "" == c.strip()
+
 def bornOneCodeFile(tplName,jsonFileName,aimFile):
-	tplContent = readFile(tplName)
-	robj = jsonObj(jsonFileName)
-	tplPre = simpleReplace(tplContent)
-	realContent = renderJson(tplPre,robj)
-	f = open(aimFile,"w")
-	f.writelines(realContent)
-	f.close()
+	configFile = tplDir + sys.argv[1] + "_config.tpl"
+	if isFileEmpty(jsonFileName) and os.path.exists(configFile):
+		configContent = readFile(configFile)
+		f = open(aimFile,"w")
+		f.writelines(configContent)
+		f.close()
+	else :
+		tplContent = readFile(tplName)
+		robj = jsonObj(jsonFileName)
+		tplPre = simpleReplace(tplContent)
+		realContent = renderJson(tplPre,robj)
+		f = open(aimFile,"w")
+		f.writelines(realContent)
+		f.close()
 
 def bornTest():
 	bornOneCodeFile('class.tpl','replace.json',"class.c")
@@ -43,3 +58,4 @@ if(__name__ == "__main__"):
 	jsonFileName = sys.argv[2]
 	aimFile = sys.argv[2]
 	bornOneCodeFile(tplFilePath,jsonFileName,jsonFileName)
+
